fix(db): widen the epoch-second columns to BIGINT
Epoch seconds do not fit MySQL's ``INT``: a stored ``reset_at`` of
``15023672358`` was fine on SQLite and aborted the account write on MySQL. The
seven epoch-second columns are ``BIGINT`` in the revisions that create them, and
``20260922_000000_widen_epoch_second_columns`` widens databases that already
exist. PostgreSQL's ``int4`` had the same latent hole, so the widening revision
applies there too.

+++ app/db/alembic/versions/20260413_000000_add_accounts_blocked_at.py
@@ -31,7 +31,7 @@ def upgrade() -> None:
         return
 
     with op.batch_alter_table("accounts") as batch_op:
-        batch_op.add_column(sa.Column("blocked_at", sa.Integer(), nullable=True))
+        batch_op.add_column(sa.Column("blocked_at", sa.BigInteger(), nullable=True))
 
 
 def downgrade() -> None:
